V1.10.1

+++ Projet_langage_C/table.h
@@ -19,7 +19,7 @@ typedef struct _civil_servant
 
 
 int index_calculator(char* name, char* surname, unsigned int base,unsigned int N);
-int power (int number, int power);
+unsigned long long power (int number, int power);
 void add(char* name, char* surname, unsigned int salary);
 void load(int number_of_servant);
 int show_salary(char* name, char* surname);
